Move specific tests to seperate modules
